+++ cloud/acp-admin-cloud-dependencies/src/main/kotlin/pers/acp/admin/common/event/ReloadDataBusEvent.kt
@@ -1,0 +1,9 @@
+package pers.acp.admin.common.event
+
+import org.springframework.cloud.bus.event.RemoteApplicationEvent
+
+/**
+ * @author zhang by 19/03/2019
+ * @since JDK 11
+ */
+class ReloadDataBusEvent(origin: String?, destination: String?, val message: String, source: Any = Object()) : RemoteApplicationEvent(source, origin, destination)