+++ doc/version_history.md
@@ -2,6 +2,9 @@
 ##### v4.1.2
 > - [Upgrade] 升级 flowable 至 6.5.0
 > - [Upgrade] Spring Boot Admin 升级至 2.2.2
+> - [Upgrade] Spring Cloud Alibaba 升级至 2.2.0.RELEASE
+> - [Upgrade] gateway 中启用 spring-cloud-loadbalancer
+> - [Upgrade] Spring Boot Admin 中启用 spring-cloud-loadbalancer
 ##### v4.1.1
 > - [Upgrade] 去除多余无用依赖
 > - [Upgrade] 修改feign客户端写法