+++ doc/version_history.md
@@ -6,6 +6,7 @@
 > - [Upgrade] gateway 中启用 spring-cloud-loadbalancer
 > - [Upgrade] Spring Boot Admin 中启用 spring-cloud-loadbalancer
 > - [Upgrade] 优化工作流图片生成
+> - [Upgrade] 路由增加元数据配置
 ##### v4.1.1
 > - [Upgrade] 去除多余无用依赖
 > - [Upgrade] 修改feign客户端写法