+++ doc/version_history.md
@@ -1,4 +1,6 @@
 ## 版本更新记录
+##### v4.1.x
+> - [Upgrade] dependencies 中增加公共总线事件对象
 ##### v4.1.2
 > - [Upgrade] 升级 flowable 至 6.5.0
 > - [Upgrade] Spring Boot Admin 升级至 2.2.2